+++ src/main/java/pluginstudies/pluginstudies/DEPRECATEDCLASSES/AxeAffixes.java
@@ -1,4 +1,4 @@
-package pluginstudies.pluginstudies.deprecated;
+package pluginstudies.pluginstudies.DEPRECATEDCLASSES;
 
 import pluginstudies.pluginstudies.Crafting.Interfaces.AffixTableAcessInterface;
