+++ src/main/java/pluginstudies/pluginstudies/DEPRECATEDCLASSES/ModifierInformation.java
@@ -1,8 +1,6 @@
-package pluginstudies.pluginstudies.deprecated;
+package pluginstudies.pluginstudies.DEPRECATEDCLASSES;
 
 import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 @Deprecated
