+++ src/main/java/pluginstudies/pluginstudies/DEPRECATEDCLASSES/StatCalculator.java
@@ -1,4 +1,4 @@
-package pluginstudies.pluginstudies.deprecated;
+package pluginstudies.pluginstudies.DEPRECATEDCLASSES;
 
 import org.bukkit.inventory.ItemStack;
 @Deprecated
